Revert " #46 sqlModel の整理"
