--- nodejs/sqlModel/index.js
@@ -1,15 +1,0 @@
-module.exports = {
-    getNowEpgData: require(__dirname + '/getNowEpgData'),
-    getChannelAndGenru: require(__dirname + '/getChannelAndGenru'),
-    getRecordedList: require(__dirname + '/getRecordedList'),
-    getRecordedKeywordList: require(__dirname + '/getRecordedKeywordList'),
-    getRecordedCategoryList: require(__dirname + '/getRecordedCategoryList'),
-    getRecordedChannelList: require(__dirname + '/getRecordedChannelList'),
-    getTranscodeId: require(__dirname + '/getTranscodeId'),
-    getRecordedId: require(__dirname + '/getRecordedId'),
-    getReservationTable: require(__dirname + '/getReservationTable'),
-    getKeywordTable: require(__dirname + '/getKeywordTable'),
-    countKeywordTable: require(__dirname + '/countKeywordTable'),
-    getKeywordTableByIDAndTransexpandTable: require(__dirname + '/getKeywordTableByIDAndTransexpandTable')
-}
-
